Move run_constructor responsibility higher up
